transcription-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.0
python-telegram-bot>=20.7
openai>=1.10.0
twilio>=8.11.0
alembic>=1.13.0
pydantic-settings>=2.1.0
requests
email-validator
bcrypt==4.0.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
cryptography==42.0.5
